chage start time

+++ src/com/foxminded/universety/Universety.java
@@ -29,8 +29,8 @@ public class Universety {
         LectureHall lectureHall3 = new LectureHall(203);
         LectureHall lectureHall4 = new LectureHall(204);
 
-        Field oneMonday = new Field(subject1, (float) 9.00 , group1, teacher1, lectureHall1);
-        Field twoMonday = new Field(subject3, (float) 10.00, group1, teacher3, lectureHall3);
+        Field oneMonday = new Field(subject1,  9 , group1, teacher1, lectureHall1);
+        Field twoMonday = new Field(subject3, 10, group1, teacher3, lectureHall3);
 
         Schedule monday = new Schedule("monday");
         Schedule tuesday = new Schedule("tuesday");
@@ -38,6 +38,7 @@ public class Universety {
         monday.addSchedule(twoMonday);
         tuesday.addSchedule(twoMonday);
         tuesday.addSchedule(oneMonday);
+        
         
         Schedule june = new Schedule("June");
